Fog/fog.py
import sys
import os
import random
import string
import re
import base64
import hashlib
import hmac
import mmap
import ctypes
import secrets
from datetime import datetime
from typing import Tuple, Optional, Dict, Any
from Crypto.Cipher import AES, ChaCha20_Poly1305
from Crypto.Random import get_random_bytes
from Crypto.Util.Padding import pad, unpad
from Crypto.Protocol.KDF import scrypt
from Crypto.Hash import BLAKE2b
import psutil
import multiprocessing
from concurrent.futures import ThreadPoolExecutor
from functools import lru_cache

# Constants for security and performance
MEMORY_LIMIT = 1024 * 1024 * 100  # 100MB
SCRYPT_N = 2**20
SCRYPT_R = 8
SCRYPT_P = 1
MAX_THREADS = multiprocessing.cpu_count()
CHUNK_SIZE = 64 * 1024  # 64KB chunks for encryption

class SecurityError(Exception):
    pass

class SecureMemory:
    """Secure memory management"""
    def __init__(self):
        self.protected_memory: Dict[int, Any] = {}
        
    def protect_memory(self, data: bytes) -> int:
        """Lock memory pages to prevent swapping"""
        addr = id(data)
        try:
            mm = mmap.mmap(-1, len(data), mmap.MAP_PRIVATE)
            mm.write(data)
            mm.flush()
            self.protected_memory[addr] = mm
            return addr
        except Exception as e:
            raise SecurityError(f"Memory protection failed: {e}")

    def unprotect_memory(self, addr: int) -> None:
        """Securely clear and unprotect memory"""
        if addr in self.protected_memory:
            mm = self.protected_memory[addr]
            try:
                mm.write(b'\x00' * mm.size())
                mm.flush()
                mm.close()
                del self.protected_memory[addr]
            except Exception:
                pass

class SecureContext:
    """Context manager for secure operations"""
    def __init__(self):
        self.secure_memory = SecureMemory()
        self.thread_pool = ThreadPoolExecutor(max_workers=MAX_THREADS)
        
    def __enter__(self):
        return self
        
    def __exit__(self, exc_type, exc_val, exc_tb):
        self.thread_pool.shutdown(wait=True)
        for addr in list(self.secure_memory.protected_memory.keys()):
            self.secure_memory.unprotect_memory(addr)

@lru_cache(maxsize=32)
def compute_key_derivation(password: str, salt: bytes) -> bytes:
    """Compute key using scrypt with caching"""
    return scrypt(
        password.encode(),
        salt,
        32,  # key length
        N=SCRYPT_N,
        r=SCRYPT_R,
        p=SCRYPT_P
    )

def generate_secure_key(password: str, salt: bytes = None) -> Tuple[bytes, bytes, bytes]:
    """Generate secure keys using scrypt"""
    if not salt:
        salt = get_random_bytes(32)
    
    with SecureContext() as ctx:
        # Derive main key
        main_key = compute_key_derivation(password, salt)
        addr = ctx.secure_memory.protect_memory(main_key)
        
        # Generate additional keys
        blake = BLAKE2b.new(key=main_key, digest_bits=256)
        encryption_key = blake.digest()
        
        blake.update(b'hmac')
        hmac_key = blake.digest()
        
        ctx.secure_memory.unprotect_memory(addr)
        return encryption_key, hmac_key, salt

def secure_encrypt_chunked(data: str, key: bytes) -> Tuple[bytes, bytes]:
    """Encrypt data in chunks for better memory usage"""
    cipher = ChaCha20_Poly1305.new(key=key)
    nonce = cipher.nonce
    
    chunks = []
    for i in range(0, len(data), CHUNK_SIZE):
        chunk = data[i:i+CHUNK_SIZE].encode()
        chunks.append(cipher.encrypt(chunk))
    
    ciphertext = b''.join(chunks)
    tag = cipher.digest()
    
    return ciphertext, nonce + tag

def verify_system_security():
    """Enhanced system security checks"""
    checks = [
        # Check for debuggers
        lambda: not any(p.name().lower() in ['debugger', 'x64dbg', 'ollydbg', 'ida', 'radare2'] 
                       for p in psutil.process_iter(['name'])),
        # Check for virtualization
        lambda: not any(v in open('/proc/cpuinfo').read().lower() 
                       for v in ['hypervisor', 'vmware', 'virtualbox']),
        # Check for sufficient memory
        lambda: psutil.virtual_memory().available >= MEMORY_LIMIT,
        # Check for suspicious modules
        lambda: not any(mod in sys.modules for mod in ['pydevd', 'ida_dbg', '_pydevd_bundle']),
        # Check for system integrity
        lambda: os.path.exists('/boot/System.map-' + os.uname().release)
    ]
    
    return all(check() for check in checks)

def advanced_obfuscation(code: str) -> str:
    """Enhanced code obfuscation with performance optimizations"""
    with SecureContext() as ctx:
        # Add memory protection
        code = f"""
import ctypes
def protect_memory():
    try:
        import psutil
        p = psutil.Process()
        for m in p.memory_maps():
            addr = int(m.addr.split('-')[0], 16)
            ctypes.memmove(addr, addr, m.size)
    except: pass

protect_memory()
{code}
"""
        # Add integrity verification
        checksum = BLAKE2b.new(digest_bits=256)
        checksum.update(code.encode())
        code = f"""
_code_hash = '{checksum.hexdigest()}'
if __import__('blake2b', fromlist=['BLAKE2b']).new(digest_bits=256).update(__import__('inspect').getsource(protect_memory).encode()).hexdigest() != _code_hash:
    raise SecurityError("Code integrity check failed")
{code}
"""
        return code

def main():
    """Enhanced main function with security context"""
    if len(sys.argv) != 3:
        print("Usage: python fog.py <input_file> <output_file>")
        return

    if not verify_system_security():
        sys.exit(1)

    with SecureContext() as ctx:
        try:
            input_file = sys.argv[1]
            output_file = sys.argv[2]

            if not os.path.isfile(input_file):
                print(f"Error: File '{input_file}' does not exist.")
                return

            with open(input_file, 'r') as f:
                code = f.read()

            # Generate a strong random password
            password = base64.b64encode(get_random_bytes(32)).decode()
            
            # Apply security measures
            code = advanced_obfuscation(code)
            wrapped_code = secure_wrapper(code, password)
            
            with open(output_file, 'w') as f:
                f.write(wrapped_code)
                
            print("Code secured and saved successfully")
            
        except Exception as e:
            print(f"Error: {str(e)}")
            sys.exit(1)
        finally:
            # Ensure cleanup
            ctypes.memset(id(password), 0, len(password))

if __name__ == "__main__":
    main()
