Fog/requirements.txt
# Requirements for fog.py
pycryptodome>=3.15.0  # For cryptographic operations (AES, ChaCha20_Poly1305)
psutil>=5.9.0         # For system monitoring and security checks
